requirements.txt
defusedxml==0.4.1
Django==1.9.8
django-allauth==0.28.0
djangorestframework==3.4.1
oauthlib==2.0.0
psycopg2==2.6.2
python3-openid==3.0.10
requests==2.11.1
requests-oauthlib==0.7.0
